fix(nfebrasil): nfse_emissao_id unsignedInteger casa nfse_emissoes.id (migrate fresh) [M]
Erro 3780 no migrate do-zero: FK nfse_eventos_canc_emi_fk falhava porque
nfse_emissao_id era unsignedBigInteger e nfse_emissoes.id é increments (unsigned
INT). FOREIGN_KEY_CHECKS=0 não bypassa type-mismatch no DDL. Alinha o tipo.
Prod inalterado (migration já aplicada — não re-roda). Destrava lane US-FIN-052.

+++ Modules/NfeBrasil/Database/Migrations/2026_05_12_120000_create_nfse_eventos_cancelamento_table.php
@@ -49,7 +49,9 @@ return new class extends Migration
         Schema::create('nfse_eventos_cancelamento', function (Blueprint $table) {
             $table->id();
             $table->unsignedInteger('business_id')->index();
-            $table->unsignedBigInteger('nfse_emissao_id');
+            // nfse_emissoes.id é `increments` (unsigned INT) — a coluna FK precisa
+            // ser unsignedInteger pra casar o tipo (senão erro 3780 no migrate fresh).
+            $table->unsignedInteger('nfse_emissao_id');
             $table->string('driver_key', 32)
                 ->comment('ABRASF_V1, ABRASF_V2.04, GINFES, IPM, TIPLAN, NFSE_GOV_BR');
             $table->text('motivo')
